+++ text.py
@@ -42,3 +42,47 @@ Regain [[2d8+5]] HP | Moderate, **Moderate Healing Potion**
 Regain [[3d8+10]] HP | Greater, **Greater Healing Potion**
 Regain [[6d8+20]] HP | Major, **Major Healing Potion**
 Regain [[8d8+30]] HP}}}''')
+
+
+def cprint(color, string, end='\n', file=sys.stdout):
+    bcolors = {
+        'HEADER': '\033[95m',
+        'OKBLUE': '\033[94m',
+        'OKCYAN': '\033[96m',
+        'OKGREEN': '\033[92m',
+        'WARNING': '\033[93m',
+        'FAIL': '\033[91m',
+        'ENDC': '\033[0m',
+        'BOLD': '\033[1m',
+        'UNDERLINE': '\033[4m'
+    }
+
+    if color.upper() in bcolors:
+        head = bcolors[color.upper()]
+    else:
+        head = f'\x1b[{color}m'
+
+    tail = '\x1b[0m'
+
+    print(head + string + tail, end=end, file=file)
+
+
+def error(message, exitcode=None):
+    cprint('0;31;40', f'ERROR: {message}', file=sys.stderr)
+
+    if exitcode:
+        sys.exit(exitcode)
+
+
+
+def bar(header=None, char='#', length=os.get_terminal_size()[0]):
+    output = ''
+
+    if header:
+        output += (length*char + '\n')
+        output += (char + header.center(length-2) + char + '\n')
+
+    output += (length*char + '\n')
+    return output
+
+
